TST: Fixed tests to work with enw format.  A couple notes:
The test does a confidence interval and test for each parameter twice, one for 'powell'and one for 'nm'.

'nm' is consistently faster than 'powell' with the total time for 'nm' taking about 30 seconds and the total time for 'powell' taking 100 seconds.  Maybe 'powell' should be removed as a fit method since the except LinAlgError in the optimization routine guards against optimizers returning an error for very unlikely values of nuisance parameters.