ENH: speed up bandwidth estimation for nonparametric.KDE
gpke() should only accept ndarrays instead of array_like, same for kernels.
They get called so often that it doesn't make sense to call asarray in them.
On a simple test asarray got called 1e5 times for one bw estimation.

Speedup of this commit ~35% on bw estimation in KDE with cv_ls for the dataset
in the tests.