Summary_col: Mechnism for default model infos and overwrites per model type
Results can add a property `default_model_infos` to specify a info_dict,
which is taken as the default for that model. Same names are merged into
one row, even if two different models are compared in one summary_col
call.

To overwrite, info_dict can nest a model-specific info_dict under the name
of the model, which would then be used. If no such key is found in the
dict, then the complete info_dict is used.

Example:
`info_dict = {"N":..., "R2": ..., "OLS":{"R2":...}}` would not show
N for OLS regression models.