Update boxplots.py
Boxplot is broken when you use np.array instead of a list of array

The issue is it does not tell you it is broken (even if the docstring is explicit about data being list of array)

For instance  : 

sm.graphics.violinplot(np.random.normal(size=[5,1000]))

will return a graph with 5 violins each corresponding to 1000 data points and 5 boxplot corresponding to 5 data points, leaving 995 X 5 data points unplotted. 
Either it should crash and tell you to use a list of array, either it should work.

I propose to ckeck if data as 2 dimensions. If so, transpose it so its representation is coherent with the violin plot.