ENH: GLS fix for very large models
GLS can now handle very large models with diagonal sigma matrices.  This was done by removing the need for a full nobs by nobs covariance when sigma is input as a scalar or vector.

Test coverage for large model with 100,000 observations.  This would require > 64GB of main memory in previous GLS code.