ENH: Silence test output
A simple patch that makes minimal changes to tests that silences almost all output.
Most fixes were simply changing options to not output (e.g. disp=False), and a few
others trapped errors.

A new class was introduced in tools/testing.py that allows general output in a test
to be silenced.  It would be preferred for code to only raise warnings and not simply
print, so that the warning could be captured instead.