Added 'erf' function to 'erf.h' in kernel_methods
The file defines static functions called _sm_erf and _sm_erfc.
If the compiler is C99 or not windows (so probably POSIX), then
these are simply C macros for 'erf' and 'erfc', otherwise I use the
Sun Microsystem implementation, which is the one used by GCC. License
and all have been kept untouched as required.