add reference output for testing and start parsing it
Added output from the R package tsDyn (produced by the R-file tsDyn_output_generator.r which is included in this commit).
tsDyn offers the choice of how deterministic terms (constants or linear trend) are treated. This makes it possible to test cases with det. terms in the coint. relation.
Parsing the output has also started but is not finished.