fix Python-2-related error in parsing
When opening the file with the forecasting results produced by JMulTi, I added an encoding argument to open. (This was necessary since the forecasting result files contain a "ü".) This worked fine in Python 3 but I noticed that Python 2 can't handle this argument. Now io.open is used instead of open and the corresponding errors in Python 2 are gone.