Weibull absorbing state survival
This model is a Weibull survival distribution with something special : survival can become infinite. It is useful for some analysis like unemployment duration where some people will never find a job : they come in "absorbing state".
The model allows truncated values to be used and is estimated by the maximum likelihood with reparameterization.
Coefficient's VAR matrix is estimated by the delta method. Optimizer used is Powell because it is the only one able to converge, the likelihood is pretty strange.
Parameters are like wikipedia's Weibull formula for k and lambda ; AS prop is the proportion on population in absorbing state.