Delete __main__ section from var_model.py.
The example there belongs elsewhere, as this module is not intended to run as a script.

Moreover, attempting to run it raises an error when it tries to run parse_lutkepohl_data and the passed path is not found.

But most of all, deleting this should marginally increase coverage, so that coveralls will no longer mark this PR as unsuccessful