Fix AttributeError by subclassing VAR
I expect `SVAR` was at some point intended to subclass `VAR`.  Subclassing lets us inherit the `select_order` method.

This commit needs to be paired with the next one (editing on GH, not sure if I can do multi-file commits with the browser editor).  The current implementation of `select_order` calls `_estimate_var`.  That will be changed to `_estimate`, which will be aliased to `_estimate_var` in `VAR`, but which points to `_estimate_svar` here.