Use `self.J-1` instead of defining `J`
in `MNLogit.hessian`.

Relates to #3725.  While not an actual bug, the current behavior is definitely misleading.  The code currently defines `J` in a way that equals `self.J - 1`.  The variable name is sufficiently distinctive that a reader (i.e. me an hour ago) can easily make the incorrect inference that `J == self.J`.  This PR just fixes that.