Cleanup and flake8 fixes for discrete_model
Weibull model had been commented-out in its entirety since at least Oct 2012.  Deleted it.

Removed duplicate and unused imports.  Arranged imports by source

Fix docstring indentation and some comments in __main__ that were not indented
This is useful because it allows for code-folding in many IDEs.

Changed MultinomialResults._get_endog_name to avoid built-in name 'all' in favor of 'use_all'.
Use pandas.util.decorators.deprecate_kwarg to keep this backward-compatible.  As a private
method, this should not matter much.

MultinomialResults._maybe_convert_ynames_int never uses , so does not need to be a method.  Moved
it out to be a private module-level function.

Add shebang