BF/NF: Aggregate data set before performing RM-Anova if needed
If the data set contains more than one observation per subject and
experimental condition, the current AnovaRM implementation will generate
incorrect results.

This commit adds an extra aggregation step before performing the
Anova. The user is free to specify an aggregation function via the
`aggregate_func` keyword argument.

We raise an exception if aggregation is needed, but not explicitly
requested.

This functionality aims to resemble the behavior of the popular R
package, afex, which features a `fun_aggregate` parameter. Quote from
the afex docs [1]:

"fun_aggregate

The function for aggregating the data before running the ANOVA if
there is more than one observation per individual and cell of the
design. The default NULL issues a warning if aggregation is necessary
and uses mean. Pass mean directly to avoid the warning."

The ez::ezANOVA() R function behaves in a similar manner.

References:
[1] https://cran.r-project.org/web/packages/afex/afex.pdf