BUG: delay evaluation of fit results in ProbPlot
Currently, graphics.gofplots.ProbPlot always calls the fit() method of
the scipy.stats distribution during initialization, and save the
return values as the fit_params attribute.  However, this is not always
necessary.

This patch delays the evaluation of fit_params by converting it to a
property.  To do this, a new (private) attribute is added to ProbPlot,
namely _unfrozen_dist that references the unfrozen scipy.stats
distribution.

Fixes #3547