BUG: normal_sample_size_one_tail: fix mult by None
Multiplying a float by a NoneType object is a syntax error in Python,
yet this operation occurs if
`statsmodel.stats.power.normal_sample_size_one_tail` is called without
specifying a `std_alternative` parameter. This commit fixes that bug
by adding a conditional in the body of the method that assigns zero to
that parameter if its value is `None`. `None` is an idiomatically
sensible default value, which is why this default is not changed.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>