ENH: normal sample size: test the None default arg
One of the bugs this pull request is fixing is that using the default
`None` value of `std_alternative` does not raise an exception (due to
attempting to muiltiply a `float` and a `NoneType`). This commit adds
a simple smoke test and doesn't check the value, because the only
point of the test is to check that an exception is *not* raised.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>