BUG: sample size: return nan if power < alpha
Raising an exception when `power < alpha` would be poor user
experience because any element satisfying this condition could halt
computation due to raising an exception. Instead of raising an
exception, this commit masks the output array using an array of the
same shape in which elements greater than zero correspond to locations
where `alpha > power`. Elements of `n1` satisfying this condition are
set to `numpy.nan`; otherwise, elements of `n1` are left unchanged
from the sample size computation. In order to ensure this masking
works for all collections of legal, broadcastable arguments,
`numpy.ones_like` is used to broadcast the `alpha > power` condition
appropriately.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>