Avoid random ordering in include_dirs lists
I'm rebuilding RPM packages to check build reproducibility
(https://docs.fedoraproject.org/en-US/reproducible-builds/), and we
got the result that rebuilds of statsmodels varies in this trivial way:
│ │          "include_dirs": [
│ │ -            "statsmodels/src",
│ │              "/usr/lib64/python3.13/site-packages/numpy/core/include",
│ │ +            "statsmodels/src",
│ │              "/usr/lib64/python3.13/site-packages/numpy/core/include"
│ │          ],
(the same pattern repeats hundreds of times).

sets are hash-ordered, not insertion ordered, so the order in
list(set(…)) is not reproducible. Sort the entries instead.