Update model.py --corrected wald test error for RegimeSwitchingmodels
Running Wald test on a MarkovRegressionResults object (Regime Switching models) throws an attribute error as below:

AttributeError: 'MarkovRegressionResults' object has no attribute 'df_resid'

Infact, the 'MarkovRegressionResults' has a 'resid' attribute for residuals. 

Therefore, added a try/except statement to the function 'wald_test' so that for result objects that have the residuals in the attribute '.resid' don't throw an attribute error. 

This corrects the issue.