ENH: Add post-hoc tests to AnovaRM results
Adds pairwise_tukeyhsd and llpairtest methods to the AnovaResults
class. This allows for direct post-hoc analysis following a repeated
measures ANOVA, simplifying the user workflow.

This change also refactors the GroupsStats class out of the sandbox
and into the central 	ools.grouputils module. The _check_data_balanced
method in AnovaRM is updated to use this new utility.

Unit tests are added for the new post-hoc methods and for the refactored
GroupsStats class to ensure correctness and reliability.